Started using RUC LSM version from WRFV3.5.1
Changed to the WRFV3.5.1 version of the RUC LSM due to errors I was receiving (AVOST IN VILKA) using the modified version that was in the WRFV3.4.1_WOF release.

A similar error still occurs in the V3.5.1 RUC LSM version if the number of tiles is large (160), but doesn't if that number is small (32).